requirements.txt
requests==2.31.0
clickhouse-driver==0.2.6
websocket-client==1.6.1
PyYAML==6.0.1
pandas==2.0.3
schedule==1.2.1
python-dateutil==2.8.2
fastapi==0.103.2
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
redis==5.0.8
